+++ .travis_commands.bash
@@ -60,7 +60,7 @@ EOF
   sudo docker run -v "$(pwd):/home/pony" -u pony:2000 --rm ponylang/ponyc-ci:centos7-llvm-3.9.1 sh -c "make LLVM_CONFIG=/opt/llvm-3.9.1/bin/llvm-config arch=x86-64 tune=generic default_pic=true use=llvm_link_static DESTDIR=ponyc.AppDir ponydir=/usr prefix= test-ci"
   sudo docker run -v "$(pwd):/home/pony" -u pony:2000 --rm ponylang/ponyc-ci:centos7-llvm-3.9.1 sh -c "make LLVM_CONFIG=/opt/llvm-3.9.1/bin/llvm-config arch=x86-64 tune=generic default_pic=true use=llvm_link_static DESTDIR=ponyc.AppDir ponydir=/usr prefix= install"
 
-  # build stdlib to ensure libraries like openssl and pcre2 get packaged in the appimage
+  # build stdlib to ensure libraries get packaged in the appimage
   sudo docker run -v "$(pwd):/home/pony" -u pony:2000 --rm ponylang/ponyc-ci:centos7-llvm-3.9.1 sh -c "./build/release/ponyc packages/stdlib"
   sudo docker run -v "$(pwd):/home/pony" -u pony:2000 --rm ponylang/ponyc-ci:centos7-llvm-3.9.1 sh -c "cp stdlib ponyc.AppDir/usr/bin"
 
@@ -176,7 +176,6 @@ ponyc-build-debs-debian(){
   #
   # To update the builder image for an existing debian version
   # * Create and upload a new version of the appropriate docker image using .ci-dockerfiles/deb-builder
-  build_deb buster
   build_deb stretch
 
   ls -la